Call out when target date has elapsed
When the target date the user has set for a book has passed, the app will now display an alert icon, change the row color to gray, and display a message instructing the user to pick a new target date. Previously, we would just say "You've read enough today to stay on track" since our calculation was returning negative days remaining in which to finish the book.