Read books from file before determining daily targets when scenePhase becomes active
Fixes an issue where the widget would update the daily targets and write them to the JSON file, but then when you opened the app it would re-calculate the targets since it thought they needed to be updated since it wasn't pulling in the changes the widgets had made to the JSON file. This now reads from the JSON when the app re-enters the foreground, thus picking up anything the widgets might've done. It also uses the more modern .onChange approach to detecting the scene phase instead of the crazy NotificationCenter nonsense I was doing before.