Fix issue with unwrapping optional, adjust book choices for lock-screen widget
Instead of explicitly unwrapping an optional when reconciling a selected book in the widget, we're now checking to make sure it exists first and just not adding it if it doesn't. Previously this would crash the widget.

Also adjusted the accessory rectangular widget to only take 1 option instead of 2 since it can only display one anyway.